Audio spawning works!
Well, you don't really spawn audio, but whatever.

For some reason there's a delay in the spawn of the sphere, but
whatever.  I'll figure that out later.  It might have to do with the
loading of the Audio file, in which case sorry but I can't do shit about
it.
Also implemented text wrap functionality for the thumbnails.
Probably some other stuff too but I forgot it.
Next up, code cleanup!