Adding terrain functionality to main canvas.
After a rather thorough fucking of the main scene via a merge, we
decided to just revert back to where we were before the Terrain merge,
and add the terrain stuff in separately by copying it over from a
separate scene.  This commit is the result of that copying, and it
appears as though all functionality has been restored.

*crosses fingers vigorously*